actually with subsampled mod2...
